requirements.txt
playwright
pytest
PyYAML
colorlog